+++ pqreflection/pqreflection.csproj
@@ -4,11 +4,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{D064263D-1601-4B2D-959F-B208DC2167EA}</ProjectGuid>
+    <ProjectGuid>{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>pqlib</RootNamespace>
-    <AssemblyName>pqlib</AssemblyName>
+    <RootNamespace>pqreflection</RootNamespace>
+    <AssemblyName>pqreflection</AssemblyName>
     <DefaultLanguage>en-US</DefaultLanguage>
     <TargetPlatformIdentifier>UAP</TargetPlatformIdentifier>
     <TargetPlatformVersion Condition=" '$(TargetPlatformVersion)' == '' ">10.0.16299.0</TargetPlatformVersion>
@@ -106,17 +106,16 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="AppStoreServices\AppUpdater.cs" />
-    <Compile Include="AppStoreServices\IAppUpdater.cs" />
+    <Compile Include="AttributeHelper.cs" />
+    <Compile Include="DisplayAttributeExtension.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="Properties\pqlib.rd.xml" />
+    <EmbeddedResource Include="Properties\pqreflection.rd.xml" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NETCore.UniversalWindowsPlatform">
-      <Version>6.0.7</Version>
+      <Version>6.0.8</Version>
     </PackageReference>
   </ItemGroup>
-  <ItemGroup />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
     <VisualStudioVersion>14.0</VisualStudioVersion>
   </PropertyGroup>
