+++ pqytparser/Resources/MimeTypeEnum.cs
@@ -1,4 +1,4 @@
-﻿using pqlib.Reflection;
+﻿using pqreflection;
 using System;
 using System.ComponentModel.DataAnnotations;
