+++ projectquail.sln
@@ -13,6 +13,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "projectquailunittest", "pro
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pqcommonregex", "pqcommonregex\pqcommonregex.csproj", "{40B78344-38C5-44C8-B2C1-40C0A5B37629}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "pqreflection", "pqreflection\pqreflection.csproj", "{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -113,6 +115,22 @@ Global
 		{40B78344-38C5-44C8-B2C1-40C0A5B37629}.Release|x64.Build.0 = Release|x64
 		{40B78344-38C5-44C8-B2C1-40C0A5B37629}.Release|x86.ActiveCfg = Release|x86
 		{40B78344-38C5-44C8-B2C1-40C0A5B37629}.Release|x86.Build.0 = Release|x86
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Debug|ARM.ActiveCfg = Debug|ARM
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Debug|ARM.Build.0 = Debug|ARM
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Debug|x64.ActiveCfg = Debug|x64
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Debug|x64.Build.0 = Debug|x64
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Debug|x86.ActiveCfg = Debug|x86
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Debug|x86.Build.0 = Debug|x86
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|ARM.ActiveCfg = Release|ARM
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|ARM.Build.0 = Release|ARM
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|x64.ActiveCfg = Release|x64
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|x64.Build.0 = Release|x64
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|x86.ActiveCfg = Release|x86
+		{7EB211B5-DC0A-48C5-9657-6E1CE654DFCF}.Release|x86.Build.0 = Release|x86
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
